pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mash-shell"
version = "0.0.13"
authors = [
  { name="voschezang", email="mark.voschezang@student.uva.nl" },
]
description = "A shell and other utils"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    'aiohttp',
    'braceexpand',
    'django',
    'dominate',
    'flask',
    'mistletoe',
    'numpy',
    'pandas',
    'PyYAML',
    'quo',
    'rich',
    'requests',
    'scipy',
    'termcolor',
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Topic :: System :: Shells",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/voschezang/mash"
"Bug Tracker" = "https://github.com/voschezang/mash/issues"
